pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "youtube-scheduler"
version = "0.1.0"
description = "Interactive CLI to upload and schedule YouTube videos with per-project isolation."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "youtube-scheduler contributors" }]
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
youtube = [
  "google-api-python-client>=2.120.0",
  "google-auth>=2.27.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
]

[project.scripts]
yt-scheduler = "youtube_scheduler.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


